move and clarify purpose of existing id wrappers
These wrappers are necessary for bridge purposes when one is mixing
(willingly or unwillingly) C++ and C-api from circumstances such as
third parties, preexisting libraries/legacy code, or integrating
otherwise with C-only projects.  There is no need to expect an
application as one or the other as the C++ library is merely a light-
wrapper around C and there is no sense in pretending otherwise.